Rename the new enforcement action DTO
