+++ src/AppServices/Enforcement/EnforcementActionCommand/CreateEnforcementActionDto.cs
@@ -2,14 +2,14 @@
 
 namespace AirWeb.AppServices.Enforcement.EnforcementActionCommand;
 
-public record CreateEnforcementAction
+public record CreateEnforcementActionDto
 {
     [DataType(DataType.MultilineText)]
     [StringLength(7000)]
     public string? Comment { get; init; }
 
     [Display(Name = "Response requested")]
-    public bool ResponseRequested { get; init; }
+    public bool ResponseRequested { get; init; } = true;
 
     public EnforcementActionType ActionType { get; init; }
 }
